// Frontend/src/app/models/order.model.ts
export interface DonHang {
    maDH: number;
    tenKH: string;
    email: string;
    diaChiGiaoHang: string;
    sdt: string;
    ghiChu: string;
    ngayDat: string;
    phuongThucThanhToan: string;
    trangThai: number;

    maND: number;
    tongTien: number;
}
export interface CTDonHang {
    maCTDH: number;
    soLuong: number;
    giaBan: number;
    maSP: number;

    tenSP: string;
    anhDaiDien: string;
}